Fix generating function migrations for functions in custom schemas.
Currently things don't quite work right if you try to generate a
function migration for a function inside a custom PostgreSQL schema. The
underlying `create_function` (and other methods) work with a
schema-specified name, however the migration file names and file naming
assumptions get confused when a schema name is in the mix.

For example, if you currently run a command like this:

```
rails generate fx:function my_schema.my_function
```

You'll end up with `db/functions/my_schema.my_function_v01.sql` and
`db/migrate/[timestamp]_create_function_my_schema.my_function.rb`
files. The extra dot in the migration filename causes issues for Rails,
as does the invalid class name it uses inside (`class
CreateFunctionMySchema.myFunction < ActiveRecord::Migration[7.0]`).

This fixes the issue by replacing dots in the file name/table name with
underscores, which allows for the normal Rails filename and class name
conventions to work. But the original name, with the dot, is still
retained and used inside the migration itself for the `create_function`
method call, so the underlying SQL for creating and rolling back
functions is correctly schema-specified.

This adds test coverage for this function generator handling
schema-specified table names. It also adds test coverage for
schema-specified table names being used with the trigger generator, but
that all worked already since for triggers the table name wasn't present
in the file name, so this wasn't an issue.